refactor: 프로젝트 구조 변경(Content 컴포넌트 추가)
