+++ data-source/db.json
@@ -1,9 +1,9 @@
 {
   "issues": [
     {
-      "labelName": "documentation",
+      "labelName": "documentation1",
       "desc": "Improvements or additions to documentation12",
-      "color": "#8BB659",
+      "color": "#C5C6F2",
       "id": 2
     },
     {
@@ -49,10 +49,35 @@
       "id": 10
     },
     {
-      "labelName": "1212",
-      "desc": "12122",
-      "color": "#88400B",
+      "labelName": "2222",
+      "desc": "33333",
+      "color": "#662744",
       "id": 11
+    },
+    {
+      "labelName": "2222",
+      "desc": "1121221",
+      "color": "#7DE446",
+      "id": 12
+    },
+    {
+      "labelName": "444444444444444444444",
+      "desc": "",
+      "color": "#FC81C1",
+      "id": 13
+    },
+    {
+      "labelName": "55555555555",
+      "desc": "333334",
+      "color": "#BD7495",
+      "id": 14
+    },
+    {
+      "labelName": "11111111111111",
+      "desc": "111111111111111111111111111",
+      "color": "#E15097",
+      "id": 15
     }
-  ]
+  ],
+  "milestone": []
 }
