+++ src/components/Label/Label.jsx
@@ -2,11 +2,11 @@ import React from 'react'
 import LabelCreateForm from './LabelCreateForm/LabelCreateForm'
 import LabelTable from './LabelTable/LabelTable'
 
-const Label = ({ isShowForm, showForm, labels, updateData }) => {
+const Label = ({ isShowForm, toggleForm, labels, updateData }) => {
 
   return (
     <>
-      <LabelCreateForm {...{ isShowForm, showForm }}/>
+      <LabelCreateForm {...{ isShowForm, toggleForm }}/>
       <LabelTable {...{ labels, updateData }}/>
     </>
   )
