+++ src/components/Label/LabelCreateForm/LabelCreateForm.jsx
@@ -1,11 +1,11 @@
-import { LabelCreateFormWrap, } from './LabelCreateForm.style'
+import {LabelCreateFormWrap,} from './LabelCreateForm.style'
 import LabelForm from '../LabelForm/LabelForm'
 
-const LabelCreateForm = ({isShowForm = false, showForm }) => {
+const LabelCreateForm = ({isShowForm = false, toggleForm }) => {
 
   return (
     <LabelCreateFormWrap  isShowForm={isShowForm}>
-      <LabelForm onClose={showForm}/>
+      <LabelForm onClose={toggleForm}/>
     </LabelCreateFormWrap>
   )
 }
